requirements.txt
py2neo>2020.0.0
neomodel
pyyaml